Migrated project to using ParcelJs
The project up until now was using a custom pipeline which was compiling
sass and elm files and was serving them. This was a temp solution until
something better and cleared could be found ( and until I got some free
time on my hands ). Hence its now migrated onto using parceljs which
seems to be doing the job!